LPS-153397 Create the Tags Subscribed section on Subscriptions page
